Extracting logic into AutoloaderUtil
